Fix tests, move package, cleanup buildscript
- Fixed functional tests
- Fixed normal tests
- Removed Android functional tests
- Cleaned up buildscript
- Cleaned up GroovyDocs to make them render in website properly
- Updated buildscript dependencies
- Disabled configuration cache (the plugin still supports it and is
tested against it)
