--- .travis.yml
@@ -1,13 +1,0 @@
-sudo: false
-dist: trusty
-
-language: groovy
-jdk:
-  - openjdk8
-  - oraclejdk8
-
-install: true
-script: ./gradlew build
-
-notifications:
-  email: false
