--- buildSrc/build.gradle
@@ -1,14 +1,0 @@
-plugins {
-    id 'java-gradle-plugin'
-}
-
-java.toolchain.languageVersion = JavaLanguageVersion.of(8)
-
-sourceSets.main.groovy.srcDirs = ['../src/main/groovy']
-
-gradlePlugin.plugins {
-    licenser {
-        id = 'net.minecraftforge.licenser'
-        implementationClass = 'org.cadixdev.gradle.licenser.Licenser'
-    }
-}
