+++ gradle.properties
@@ -1,4 +1,27 @@
+# Build Cache
+# https://docs.gradle.org/current/userguide/build_cache.html
 org.gradle.caching=true
+
+# Parallel Project Execution
+# https://docs.gradle.org/current/userguide/performance.html#parallel_execution
 org.gradle.parallel=true
+
+# Configure On Demand
+# https://docs.gradle.org/current/userguide/multi_project_configuration_and_execution.html
 org.gradle.configureondemand=true
-org.gradle.configuration-cache=true
+
+# Configuration Cache
+# TODO [Licenser][Gradle9] Re-enable in Gradle 9 (if updated)
+# https://docs.gradle.org/current/userguide/configuration_cache.html
+#org.gradle.configuration-cache=true
+
+# Parallel configuration caching
+# TODO [Licenser][Gradle9] Re-enable in Gradle 9 (if updated)
+# https://docs.gradle.org/current/userguide/configuration_cache.html#config_cache:usage:parallel
+#org.gradle.configuration-cache.parallel=true
+
+# Ignoring problems with configuration cache
+# Fixes issues with core plugins (publish, idea, eclipse, etc.) that are not compatible with the configuration cache
+# TODO [Licenser][Gradle9] Re-enable in Gradle 9 (if updated)
+# https://docs.gradle.org/current/userguide/configuration_cache.html#config_cache:usage:ignore_problems
+#org.gradle.configuration-cache.problems=warn
