+++ settings.gradle
@@ -1,5 +1,16 @@
 plugins {
-    id 'org.gradle.toolchains.foojay-resolver-convention' version '0.7.0'
+    id 'org.gradle.toolchains.foojay-resolver-convention' version '0.9.0'
 }
 
 rootProject.name = 'licenser'
+
+//@formatter:off
+dependencyResolutionManagement.versionCatalogs.register('libs') {
+    // Testing
+    version 'spock', '2.3-groovy-3.0'
+    library 'spock-core',   'org.spockframework', 'spock-core'   versionRef 'spock'
+    library 'spock-junit4', 'org.spockframework', 'spock-junit4' versionRef 'spock'
+    bundle  'spock', ['spock-core', 'spock-junit4']
+    library 'junit',        'junit',              'junit'      version '4.13.2'
+}
+//@formatter:on
